Modif type c

+++ src/ch/epfl/alpano/gui/ChannelPainter.java
@@ -16,8 +16,22 @@ import ch.epfl.alpano.Panorama;
  */
 @FunctionalInterface
 public interface ChannelPainter {
+    
+    
+    /**
+     * Methode abstraite retournant la valeur du canal en ce point
+     * @param x int coordonnee entiere x d'un point 
+     * @param y int coordonnee entiere y d'un point
+     * @return float valeur du canal en ce point
+     */
     float valueAt(int x, int y);
     
+    /**
+     * Methode static pour calculer la difference de distance entre un point et le plus lointain de ses voisins 
+     * @param panorama Panorama à parcourir
+     * @return ChannelPainter dont la valeur pour un point est la difference de distance entre le plus lointain des voisins 
+     *          et le point en question 
+     */
     static ChannelPainter maxDistanceToNeighbors(Panorama panorama){
         
         
@@ -41,34 +55,64 @@ public interface ChannelPainter {
         
     }
     
-    default ChannelPainter add(int c){
+    /**
+     * Methode par default pour ajouter la valeur produite par le peintre par une constante passee en argument
+     * @param c int constante a ajouter a la valeur
+     * @return ChannelPainter avec 
+     */
+    default ChannelPainter add(float c){
         return (x, y) -> valueAt(x, y) + c;
     }
     
-    default ChannelPainter sub(int c){
+    /**
+     * @param c
+     * @return
+     */
+    default ChannelPainter sub(float c){
         return (x, y) -> valueAt(x, y) - c;
     }
     
-    default ChannelPainter mul(int c){
+    /**
+     * @param c
+     * @return
+     */
+    default ChannelPainter mul(float c){
         return (x, y) -> valueAt(x, y) * c;
     }
     
-    default ChannelPainter div(int c){
+    /**
+     * @param c
+     * @return
+     */
+    default ChannelPainter div(float c){
         return (x, y) -> valueAt(x, y) / c;        
     }
     
+    /**
+     * @param op
+     * @return
+     */
     default ChannelPainter map(DoubleUnaryOperator op){
         return (x, y) -> (float)op.applyAsDouble(valueAt(x, y));
     }
     
+    /**
+     * @return
+     */
     default ChannelPainter inverted(){
         return (x, y) -> 1 - valueAt(x, y);
     }
     
+    /**
+     * @return
+     */
     default ChannelPainter clamped(){
         return (x, y) -> max(0 , min(valueAt(x, y), 1));
     }
     
+    /**
+     * @return
+     */
     default ChannelPainter cycling(){
         return (x, y) -> (float)floorMod(valueAt(x, y), 1);
     }
